Add preprocessor bash script for syncing documentation files to wiki
Change documentation file names to be more user-friendly by removing the special characters.
Thus add refactoring within the script that renames the files (to be compatible with GItHub wiki) before pushing to wiki.
Also add preprocessing within the script that automatically replaces pre-defined macros within documentation files with correct and updated links.

This commit makes bi-directional syncing much more tedious, so it has been dropped. Only changes in the docs folder in the main repo will be pushed to the wiki. Changes directly to the wiki repo will no longer be persisted.
The contribution guidelines will need to be significantly udpated after this commit.
The script could also be converted into CMake instead of bash if need be,